refactor(codegen/nodejs): rename dir from ts to nodejs
