+++ codegen/generator/nodejs/templates/functions.go
