+++ codegen/generator/nodejs/templates/functions_test.go
