+++ codegen/generator/nodejs/templates/src/api.ts.tmpl
