+++ codegen/generator/nodejs/templates/src/arg.ts.tmpl
