+++ codegen/generator/nodejs/templates/src/arg_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"text/template"
 
-	"github.com/dagger/dagger/codegen/generator/ts/templates"
+	"github.com/dagger/dagger/codegen/generator/nodejs/templates"
 	"github.com/dagger/dagger/codegen/introspection"
 	"github.com/stretchr/testify/require"
 )
