+++ codegen/generator/nodejs/templates/src/args.ts.tmpl
