+++ codegen/generator/nodejs/templates/src/field_comment.ts.tmpl
