+++ codegen/generator/nodejs/templates/src/full_test.go
