+++ codegen/generator/nodejs/templates/src/header.ts.tmpl
