+++ codegen/generator/nodejs/templates/src/input_args.ts.tmpl
