+++ codegen/generator/nodejs/templates/src/input_args_test.go
