+++ codegen/generator/nodejs/templates/src/method.ts.tmpl
