+++ codegen/generator/nodejs/templates/src/method_solve.ts.tmpl
