+++ codegen/generator/nodejs/templates/src/object.ts.tmpl
