+++ codegen/generator/nodejs/templates/src/object_comment.ts.tmpl
