+++ codegen/generator/nodejs/templates/src/object_test.go
