+++ codegen/generator/nodejs/templates/src/objects.ts.tmpl
