+++ codegen/generator/nodejs/templates/src/objects_test.go
