+++ codegen/generator/nodejs/templates/src/return.ts.tmpl
