+++ codegen/generator/nodejs/templates/src/return_solve.ts.tmpl
