+++ codegen/generator/nodejs/templates/src/testdata/want-test-full.ts
