+++ codegen/generator/nodejs/templates/src/type.ts.tmpl
