+++ codegen/generator/nodejs/templates/src/types.ts.tmpl
